requirements.txt
pandas
llama-cpp-python
tqdm
luigi
openai
boto3